feat: 로그파일 남기도록 변경
한정된 용량에서 로그파일 생성되도록 변경
프로그램.ico 변경
버튼크기 변경

+++ jGauge/jGaugeDlg.h
@@ -83,7 +83,7 @@ public:
 	void DrawEdge(bool overRect=false);
 	void DrawDiffPixels();
 	void DrawInfomation();
-	double GetMM(int pixel);
+	double GetMM(double pixel);
 	bool LineIsNull(Line line);
 	void testFunc();
 	afx_msg void OnBnClickedBtnRoi2();
